implement login signup feature with spring security using JWT Token.

+++ src/main/java/com/sourabh/projects/airbnbcloneapp/service/RoomServiceImpl.java
@@ -3,7 +3,10 @@ package com.sourabh.projects.airbnbcloneapp.service;
 import com.sourabh.projects.airbnbcloneapp.dto.RoomDto;
 import com.sourabh.projects.airbnbcloneapp.entity.Hotel;
 import com.sourabh.projects.airbnbcloneapp.entity.Room;
+import com.sourabh.projects.airbnbcloneapp.entity.User;
 import com.sourabh.projects.airbnbcloneapp.exception.ResourceNotFoundException;
+import com.sourabh.projects.airbnbcloneapp.exception.UnAuthorisedException;
+import com.sourabh.projects.airbnbcloneapp.repository.BookingRepository;
 import com.sourabh.projects.airbnbcloneapp.repository.HotelRepository;
 import com.sourabh.projects.airbnbcloneapp.repository.InventoryRepository;
 import com.sourabh.projects.airbnbcloneapp.repository.RoomRepository;
@@ -12,6 +15,7 @@ import lombok.Data;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.modelmapper.ModelMapper;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -29,6 +33,7 @@ public class RoomServiceImpl implements RoomService {
     private final HotelRepository hotelRepository;
     private final ModelMapper modelMapper;
     private final InventoryService inventoryService;
+    private final BookingRepository bookingRepository;
 
     @Override
     public RoomDto createNewRoom(Long hotelId, RoomDto roomDto) {
@@ -36,6 +41,12 @@ public class RoomServiceImpl implements RoomService {
         Hotel hotel = hotelRepository
                 .findById(hotelId)
                 .orElseThrow(() -> new ResourceNotFoundException("Hotel not find with ID:" +id));
+
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(hotel.getOwner())){
+            throw new UnAuthorisedException("This user is not the owner of this hotel with id: " + hotelId);
+        }
+
         Room room = modelMapper.map(roomDto, Room.class);
         room.setHotel(hotel);
         room = roomRepository.save(room);
@@ -72,8 +83,18 @@ public class RoomServiceImpl implements RoomService {
     @Override
     public void deleteRoomById(Long roomId) {
 
+
+
         log.info("Deleting room with roomId {}", roomId);
         Room room = roomRepository.findById(roomId).orElseThrow(() -> new ResourceNotFoundException("Room not found with ID:" +roomId));
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(room.getHotel().getOwner())){
+            throw new UnAuthorisedException("This user is not the owner of this room with id: " + roomId);
+        }
+        if(bookingRepository.existsByRoomId(roomId))
+        {
+            throw new IllegalStateException("Cannot delete room with active booking for this room");
+        }
         inventoryService.deleteAllInventories(room);
         roomRepository.deleteById(roomId);
